Inverse dasymetric mapping

This function performs inverse dasymetric mapping on plot data. It selects plots based on given criteria, optionally aggregates them, and calculates forest fraction and Above Ground Biomass (AGB) data for each plot or cell.

Usage

invDasymetry(
  plot_data = NULL,
  clmn = "ZONE",
  value = "Europe",
  aggr = NULL,
  minPlots = 1,
  weighted_mean = FALSE,
  is_poly = TRUE,
  parallel = FALSE,
  n_cores = parallel::detectCores() - 1,
  dataset = "custom",
  agb_raster_path = NULL,
  forest_mask_path = NULL,
  threshold = 0,
  esacci_biomass_year = "latest",
  esacci_biomass_version = "latest",
  esacci_folder = "data/ESACCI-BIOMASS",
  gedi_l4b_folder = "data/GEDI_L4B/",
  gedi_l4b_band = "MU",
  gedi_l4b_resolution = 0.001,
  timeout = 600
)

Arguments

plot_data

data.frame, Plot dataset containing required columns.

clmn

character, Column name for plot selection.

value

character, Value to select in the specified column.

aggr

numeric, Aggregation factor in degrees. NULL for no aggregation.

minPlots

integer, Minimum number of plots per aggregated cell.

weighted_mean

logical, Whether to use weighted mean for calculations.

is_poly

logical, Whether input plots are polygons.

parallel

logical, Enable parallel processing. Default is FALSE.

n_cores

numeric, Number of cores to use for parallel processing.

dataset

character, Dataset for AGB estimation: "custom", "esacci", or "gedi". Default is "custom".

agb_raster_path

character, File path to the custom AGB raster.

forest_mask_path

character, File path to the forest mask raster.

threshold

numeric, Threshold (0-100) for tree cover calculation and forest masking (e.g. 0 or 10).

esacci_biomass_year

The ESA CCI BIOMASS AGB tiles year to use. Use either 2010, 2015, 2016, 2017, 2018, 2019, 2020, 2021 or "latest" (default).

esacci_biomass_version

The ESA CCI BIOMASS AGB tiles version to use. Use either "v2.0", "v3.0", "v4.0", "v5.0", "v5.01" or "latest" (default).

esacci_folder

Directory to save downloaded ESA CCI BIOMASS AGB files. Default is the relative path "data/ESACCI-BIOMASS".

gedi_l4b_folder

Character, the folder to save the downloaded GeoTIFF file. Default is "data/GEDI_L4B/".

gedi_l4b_band

Character, the band to filter for. See options in the Details section below. Default is "MU".

gedi_l4b_resolution

Numeric, the spatial resolution of the processed output GeoTIFF in degrees. The native resolution of the GEDI L4B gridded dataset is 1 km, approximately 0.001 degrees at the equator. Default is 0.001.

timeout

Number of seconds for reaching file download timeout.

Value

A data frame with the following columns:

plotAGB_threshold

AGB values for the given forest threshold (e.g., plotAGB_10 if threshold=10)

tfPlotAGB

Tree-filtered plot AGB (only when not aggregated)

orgPlotAGB

Original plot AGB

mapAGB

AGB from map sampling

SIZE_HA

Plot size in hectares

x

X-coordinate of plot

y

Y-coordinate of plot

n

Number of plots (only if aggregated)

Details

The function performs the following steps:

Selects plots based on the specified criteria.

Optionally aggregates plots if aggr is not NULL.

Calculates forest fraction and AGB for each plot or cell.

Samples AGB from custom raster or downloads and uses ESA CCI BIOMASS or GEDI L4B data.

Note

Ensure all required columns are present in the input plot data.

For parallel processing, adjust n_cores based on your system's capabilities.

Large datasets may require significant processing time and memory.

Examples

if (FALSE) { # \dontrun{
# Create a sample dataset of 10 plots:
set.seed(42)
sampled_plots <- plots[sample(nrow(plots), 10), ]
plot_data <- Deforested(sampled_plots, gfc_folder = test_dir,  map_year = test_year)
plot_data <- BiomePair(sampled_plots$non_deforested_plots)
plot_data <- TempApply(plot_data, test_year)

result <- invDasymetry(plot_data = plot_data,
                       clmn = "ZONE",
                       value = "Europe",
                       aggr = 0.1,
                       parallel = TRUE,
                       dataset = "esacci")
} # }